+++ packages/flutter_chat/lib/flutter_chat.dart
@@ -1,7 +1,5 @@
-// ignore_for_file: prefer_double_quotes
-
 // Core
-export 'package:chat_repository_interface/chat_repository_interface.dart';
+export "package:chat_repository_interface/chat_repository_interface.dart";
 
 // User story
 export "package:flutter_chat/src/flutter_chat_entry_widget.dart";
@@ -11,7 +9,7 @@ export "package:flutter_chat/src/flutter_chat_navigator_userstory.dart";
 export "src/config/chat_builders.dart";
 export "src/config/chat_options.dart";
 export "src/config/chat_translations.dart";
-export 'src/config/screen_types.dart';
+export "src/config/screen_types.dart";
 
 // Screens
 export "src/screens/chat_detail_screen.dart";
@@ -21,3 +19,7 @@ export "src/screens/creation/new_chat_screen.dart";
 export "src/screens/creation/new_group_chat_overview.dart";
 export "src/screens/creation/new_group_chat_screen.dart";
 export "src/services/date_formatter.dart";
+
+// Utils
+export "src/util/scope.dart";
+export "src/util/utils.dart";