+++ packages/flutter_chat/lib/src/util/utils.dart
@@ -1,0 +1,1 @@
+// add generic utils that are used in the package